Add support for schema aliasing
This adds a new file which lists the schemas we want to alias.
This does not handle versions; the aliased namespace/doctype
are assumed to have the same versions as the source schemas.

This also includes a validation schema for the aliasing file,
and tests that the file is valid.