Accumulate http response as bytestring
After https://github.com/mozilla-services/mozilla-pipeline-schemas/pull/579
was merged, schema generation was raising an error when processing the
heartbeat schema.

I believe what happened is that one of our 1024 byte chunks happened to land
within a multibyte character. We were decoding each chunk to text, which wasn't
tolerant to the possibility of a chunk ending within a multibyte character.

With this change, I was able to successfully run on the current mps content.