Correctly calculate schema size for json columns: 1
This uses the fact that we tag `object` metrics with `format=json` in
the schema, which is ... sort of correct.
Technically `format` only applies to `type: string` fields in
jsonschema, but nothing blows up if you just set it.

We need _some_ signal for json columns.
The jsonschema-transpiler actually relies on schema configuration to
know when to output a JSON column.

Yes, that's now two different ways on how to do it.
Maybe this SHOULD use what's configured in `mozPipelineMetadata`
instead?